componentize drawer
