+++ app/components/address-mapper.tsx
@@ -1,16 +1,9 @@
 "use client";
 
-import { useCallback, useEffect, useRef, useState } from "react";
+import { useCallback, useEffect, useState } from "react";
 import { useRouter, usePathname, useSearchParams } from "next/navigation";
-import {
-  chakra,
-  useDisclosure,
-  IconButton,
-  Drawer,
-  Portal,
-  Box,
-} from "@chakra-ui/react";
-import { FaAngleLeft, FaAngleRight } from "react-icons/fa";
+import { Box } from "@chakra-ui/react";
+
 import { toaster } from "@/components/ui/toaster";
 import Map from "./map";
 import ReportHazards from "./report-hazards";
@@ -18,13 +11,12 @@ import { FeatureCollection, Geometry } from "geojson";
 import HomeHeader from "./home-header";
 import { useHazardDataFetcher } from "../hooks/useHazardDataFetcher";
 import SearchBar from "./search-bar";
-import AlertInfo from "@/components/ui/alert-info";
+import SHDrawer from "./drawer";
+import AlertInfo from "./ui/alert-info";
 
 const defaultCoords = [-122.4194, 37.7949];
 
 const toggledStatesDefaults = [true, true, true];
-const AngleLeft = chakra(FaAngleLeft);
-const AngleRight = chakra(FaAngleRight);
 
 interface AddressMapperProps {
   softStoryData: FeatureCollection<Geometry>;
@@ -57,18 +49,6 @@ const AddressMapper: React.FC<AddressMapperProps> = ({
 }) => {
   const router = useRouter();
   const pathname = usePathname();
-
-  // Drawer
-  const { open, onOpen, onClose } = useDisclosure({ defaultOpen: true });
-  const drawerContainerRef = useRef<HTMLDivElement>(null);
-  const [drawerContainer, setDrawerContainer] = useState<HTMLDivElement | null>(
-    null
-  );
-
-  const setContainer = useCallback((node: HTMLDivElement | null) => {
-    drawerContainerRef.current = node;
-    setDrawerContainer(node);
-  }, []);
 
   // Search Box
   const searchParams = useSearchParams();
@@ -225,112 +205,23 @@ const AddressMapper: React.FC<AddressMapperProps> = ({
           onSearchChange={handleSearchChange}
         />
       </HomeHeader>
-      <Box w="full" h="full" m="auto" position="relative" ref={setContainer}>
+      <Box w="full" h="full" m="auto" position="relative">
         <Box h="full" overflow="hidden">
-          {/* dummy drawer, closed */}
-          {drawerContainer && !open && (
-            <Portal container={drawerContainerRef}>
-              <Box
-                position="absolute"
-                zIndex="overlay"
-                top={{ base: "auto", md: "0" }}
-                left="0"
-                bottom="0"
-                right={{ base: "0", md: "auto" }}
-                w={{ base: "auto", md: "5" }}
-                h={{ base: "5", md: "auto" }}
-                backgroundColor="white"
-              >
-                <Box
-                  onClick={onOpen}
-                  asChild
-                  position="absolute"
-                  // Mobile: center horizontally at bottom.
-                  left={{ base: "0", md: "0" }}
-                  right={{ base: "0", md: "auto" }}
-                  bottom={{ base: "0", md: "auto" }}
-                  // Desktop: vertically center relative to container.
-                  top={{ base: "auto", md: "1/2" }}
-                  w={{ base: "fit", md: "auto" }}
-                  mx={{ base: "auto", md: "0" }}
-                  transform={{ base: "none", md: "translateY(-50%)" }}
-                >
-                  <IconButton variant="subtle" rounded="full" size="md">
-                    <AngleRight rotate={{ base: "270deg", md: "0deg" }} />
-                  </IconButton>
-                </Box>
-              </Box>
-            </Portal>
-          )}
-          {/* actual drawer, open */}
-          {drawerContainer && (
-            <Drawer.Root
-              placement={{ mdDown: "bottom", md: "start" }}
-              open={open}
-              onOpenChange={(details) => {
-                if (!details.open) onClose();
-              }}
-              modal={false}
-              closeOnInteractOutside={false}
-              lazyMount
-            >
-              <Portal container={drawerContainerRef}>
-                <Drawer.Positioner
-                  h="full"
-                  w="full"
-                  position="absolute"
-                  pointerEvents="none"
-                >
-                  <Drawer.Content
-                    // NOTE: the following props are used because the `size` prop values of `Drawer.Root` are too limited (and do not directly correspond to the theme `sizes` tokens)
-                    w={{ base: "full", md: "sm" }}
-                    maxW={{ base: "full", md: "sm" }}
-                    h={{ base: "1/2", md: "full" }}
-                    maxH={{ base: "1/2", md: "full" }}
-                    pointerEvents="auto"
-                    css={{
-                      "&[data-state='open']": { animationName: "none" }, // prevent slide-in animation (`skipAnimationOnMount` on `Drawer.Root` doesn't appear to work)
-                    }}
-                  >
-                    <Drawer.CloseTrigger
-                      onClick={onClose}
-                      asChild
-                      position="absolute"
-                      // Mobile: centered above drawer edge.
-                      // Desktop: right edge, vertically centered.
-                      left={{ base: "0", md: "auto" }}
-                      right={{ base: "0", md: "-5" }}
-                      top={{ base: "-5", md: "1/2" }}
-                      w={{ base: "fit", md: "auto" }}
-                      mx={{ base: "auto", md: "0" }}
-                      transform={{ base: "none", md: "translateY(-50%)" }}
-                    >
-                      <IconButton variant="subtle" rounded="full" size="md">
-                        <AngleLeft rotate={{ base: "270deg", md: "0deg" }} />
-                      </IconButton>
-                    </Drawer.CloseTrigger>
-                    <Drawer.Header>
-                      <Drawer.Title>Risk Layers</Drawer.Title>
-                    </Drawer.Header>
-                    <Drawer.Body>
-                      <ReportHazards
-                        addressHazardData={displayData}
-                        isHazardDataLoading={isHazardDataLoading}
-                        toggledStates={toggledStates}
-                        setToggledStates={setToggledStates}
-                        setLayerToggleObj={setLayerToggleObj}
-                        isInDrawer={true}
-                      />
-                    </Drawer.Body>
-                    <Drawer.Footer>
-                      <AlertInfo message="72% chance of major Bay Area earthquake in the next 30 years" />
-                    </Drawer.Footer>
-                  </Drawer.Content>
-                </Drawer.Positioner>
-              </Portal>
-            </Drawer.Root>
-          )}
-
+          <SHDrawer
+            title="Risk Layers"
+            footerText={
+              <AlertInfo message="72% chance of major Bay Area earthquake in the next 30 years" />
+            }
+          >
+            <ReportHazards
+              addressHazardData={displayData}
+              isHazardDataLoading={isHazardDataLoading}
+              toggledStates={toggledStates}
+              setToggledStates={setToggledStates}
+              setLayerToggleObj={setLayerToggleObj}
+              isInDrawer={true}
+            />
+          </SHDrawer>
           <Map
             lon={lon}
             lat={lat}
